// src/SeoToolkit.Umbraco.SiteAudit/assets/src/api/types.gen.ts
export type SiteAuditCheckViewModel = {
    id: number;
    name?: (string) | null;
    description?: (string) | null;
    errorMessage?: (string) | null;
};

export type SiteAuditCreateConfigViewModel = {
    checks?: Array<(SiteAuditCheckViewModel)> | null;
    minimumDelayBetweenRequest: number;
    allowMinimumDelayBetweenRequestSetting: boolean;
};

export type SiteAuditDetailViewModel = {
    id: number;
    name?: (string) | null;
    maxPagesToCrawl?: (number) | null;
    totalPagesFound: number;
    progress: number;
    status?: (string) | null;
    checks?: Array<(SiteAuditCheckViewModel)> | null;
    pagesCrawled?: Array<(SiteAuditPageDetailViewModel)> | null;
};

export type SiteAuditOverviewViewModel = {
    id: number;
    name?: (string) | null;
    createdDate?: (string) | null;
    status?: (string) | null;
};

export type SiteAuditPageDetailViewModel = {
    url?: (string) | null;
    statusCode: number;
    results?: Array<(SiteAuditResultDetailViewModel)> | null;
};

export type SiteAuditResultDetailViewModel = {
    message?: (string) | null;
    checkId: number;
    isError: boolean;
    isWarning: boolean;
};